Move AbstractPage into Contracts namespace
