+++ src/Models/MarkdownDocument.php
@@ -3,6 +3,7 @@
 namespace Hyde\Framework\Models;
 
 use Hyde\Framework\Concerns\HasDynamicTitle;
+use Hyde\Framework\Contracts\AbstractPage;
 
 /**
  * The base class for all Markdown-based Page Models.
